+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # CombatInsights
+1.1.0
+    - Changes for Update 44
+    - Fixed issues around ability id changes
+    - Reworked Maelstrom 2H
+    - Grave Lord's Sacrifice support (necro mag BB morph)
+    - Follow up passive
+    - Scribing: damage skills, magic Knife debuff, banner buffs
 1.0.5
     - Fixed crash caused by Hawk Eye passive (thx EvilDave)
     - Sluthrug’s Hunger support